Filters for mf2 for main content for Post Kinds includes
Trying to include as many mf2 classes as I could statically, using Post Kinds plugin introduces quirks in syndications of content to other sites and apps since the Post Kinds plugin is using 'the_content()' as a hook for the Post Kinds content.

Now there are filters to wrap the post content so that the Post Kinds includes are not included in the same mf2 wrapped content.

Also fixed meta entry section container not to wrap the anchor, which was partially breaking the template.