rel="author" in head
For single author sites, rel="author" is now added as a link element in the <head>. This is in contrast to having rel="author" inside of a byline on every blog post or page. Indiewebify.me will complain about the lack of an author defined in each blog post but this appears as a good alternative to keep markup clutter down.